+++ geotrek/land/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 from geotrek.common.mixins.actions import MergeActionMixin
-from .models import PhysicalType, LandType
+from .models import PhysicalType, LandType, CirculationType
 
 
 class PhysicalTypeAdmin(MergeActionMixin, admin.ModelAdmin):
@@ -18,5 +18,13 @@ class LandTypeAdmin(MergeActionMixin, admin.ModelAdmin):
     merge_field = "name"
 
 
+class CirculationTypeAdmin(MergeActionMixin, admin.ModelAdmin):
+    list_display = ('name', 'structure', )
+    search_fields = ('name', 'structure')
+    list_filter = ('structure',)
+    merge_field = "name"
+
+
 admin.site.register(PhysicalType, PhysicalTypeAdmin)
 admin.site.register(LandType, LandTypeAdmin)
+admin.site.register(CirculationType, CirculationTypeAdmin)