+++ geotrek/land/templates/land/sql/post_10_views.sql
@@ -32,3 +32,10 @@ CREATE VIEW {{ schema_geotrek }}.v_workmanagements AS (
 	WHERE f.topo_object_id = e.id AND f.organization_id = b.id
 	AND deleted = FALSE
 );
+
+CREATE VIEW {{ schema_geotrek }}.v_circulations AS (
+	SELECT e.*, b.structure_id, b.name
+	FROM core_topology AS e, land_circulationedge AS f, land_circulationtype AS b
+	WHERE f.topo_object_id = e.id AND f.circulation_type_id = b.id
+	AND deleted = FALSE
+);