+++ geotrek/land/tests/factories.py
@@ -35,6 +35,20 @@ class LandEdgeFactory(TopologyFactory):
     land_type = factory.SubFactory(LandTypeFactory)
 
 
+class CirculationTypeFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = models.CirculationType
+
+    name = factory.Sequence(lambda n: "CirculationType %s" % n)
+
+
+class CirculationEdgeFactory(TopologyFactory):
+    class Meta:
+        model = models.CirculationEdge
+
+    circulation_type = factory.SubFactory(CirculationTypeFactory)
+
+
 class CompetenceEdgeFactory(TopologyFactory):
     class Meta:
         model = models.CompetenceEdge