+++ geotrek/land/tests/test_models.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 
 from geotrek.authent.models import Structure
-from geotrek.land.models import LandType, PhysicalType
+from geotrek.land.models import LandType, PhysicalType, CirculationType
 
 
 class TestModelLand(TestCase):
@@ -22,3 +22,12 @@ class TestModelLand(TestCase):
         structure = Structure.objects.create(name="Structure_1")
         pt = LandType.objects.create(name='LandType_1', structure=structure)
         self.assertEqual(str(pt), 'LandType_1 (Structure_1)')
+
+    def test_circulationtype_value_no_structure(self):
+        pt = CirculationType.objects.create(name='CirculationType_1')
+        self.assertEqual(str(pt), 'CirculationType_1')
+
+    def test_circulationtype_value_default_structure(self):
+        structure = Structure.objects.create(name="Structure_1")
+        pt = CirculationType.objects.create(name='CirculationType_1', structure=structure)
+        self.assertEqual(str(pt), 'CirculationType_1 (Structure_1)')